Renamed tools/{gstreamer-completion => gstreamer-completion-0.10}
To allow it to be installed alongside gstreamer-completion-1.0.

Note that GStreamer 1.0 only installs gst-launch-1.0 and
gst-inspect-1.0, whereas GStreamer 0.10 installs gst-launch,
gst-launch-0.10, gst-inspect and gst-inspect-0.10.

Becase this renames and modifies the file at the same time, to see the
actual changes use the "-M" flag to "git show", "git diff", or
"git log -p".